+++ NEWS.md
@@ -2,6 +2,11 @@
 
 We started keeping track of changes in the `NEWS.md` file after version 0.1.0.
 
+# PlantGeomTurtle 1.0.2 (2026-06-17)
+
+* Add `Ellipsoid!` and `Ellipsoid` turtle constructors to generate solid ellipsoids
+  in front of the turtle. Requires PlantGeomPrimitives 1.0.2.
+
 # PlantGeomTurtle 1.0.0 (2026-06-10)
 
 No actual changes. We move to version 1.0.0 because: